requirements.txt
google-generativeai>=0.4.0
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0
pydub>=0.25.1
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.97.0
schedule>=1.2.0
numpy>=1.24.0
